pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Clean Architecture Discord Bot Template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "discord-py>=2.5.2",
    "fastapi>=0.124.4",
    "flow-res>=0.1.0",
    "injector>=0.22.0",
    "python-dotenv>=1.2.1",
    "python-ulid>=3.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.38.0",
]

[project.scripts]
start-bot = "app.bot.__main__:main"
start-api = "app.api.__main__:start"

[dependency-groups]
dev = [
    "anyio>=4.11.0",
    "pre-commit>=4.5.0",
    "pyright>=1.1.407",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.14.6",
]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_empty = true
fail_under = 75
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
include = ["app", "tests" , "alembic"]
exclude = [
    "**/__pycache__",
    ".venv",
]
reportMissingTypeStubs = false
reportPrivateUsage = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_default_fixture_loop_scope = "function"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
